Refactor: split proc macro tests into multiple files
commit-id:b0254722
